+++ src/intrinsics/atomic.rs
