+++ src/shims/mod.rs
@@ -5,7 +5,6 @@ mod backtrace;
 #[cfg(target_os = "linux")]
 pub mod ffi_support;
 pub mod foreign_items;
-pub mod intrinsics;
 pub mod unix;
 pub mod windows;
 mod x86;
