+++ tests/pass/intrinsics/integer.rs
