+++ tests/pass/intrinsics/intrinsics.rs
