+++ tests/pass/intrinsics/portable-simd-ptrs.rs
