+++ tests/pass/intrinsics/portable-simd.rs
