+++ tests/pass/intrinsics/volatile.rs
