+++ tests/pass/shims/x86/intrinsics-x86-avx.rs
