+++ tests/pass/shims/x86/intrinsics-x86-avx2.rs
