+++ tests/pass/shims/x86/intrinsics-x86-pause-without-sse2.rs
