+++ tests/pass/shims/x86/intrinsics-x86-sse.rs
